remove invite user button from megamenu, clean up and consolidate usage of util functions across top nav and command palette
