+++ .github/CODEOWNERS
@@ -735,7 +735,6 @@ playwright.storybook.config.ts @grafana/grafana-frontend-platform
 /public/app/core/components/GraphNG/ @grafana/dataviz-squad
 /public/app/core/components/help/ @grafana/grafana-search-navigate-organise
 /public/app/core/components/Indent/ @grafana/grafana-search-navigate-organise
-/public/app/core/components/InviteUserButton/ @grafana/sharing-squad
 /public/app/core/components/Layers/ @grafana/dataviz-squad
 /public/app/core/components/Login/ @grafana/grafana-search-navigate-organise
 /public/app/core/components/NativeScrollbar.tsx @grafana/grafana-search-navigate-organise
