+++ public/app/core/components/AppChrome/MegaMenu/MegaMenu.tsx
@@ -13,9 +13,6 @@ import { setBookmark } from 'app/core/reducers/navBarTree';
 import { usePatchUserPreferencesMutation } from 'app/features/preferences/api/index';
 import { useDispatch, useSelector } from 'app/types/store';
 
-import { InviteUserButton } from '../../InviteUserButton/InviteUserButton';
-import { shouldRenderInviteUserButton } from '../../InviteUserButton/utils';
-
 import { MegaMenuHeader } from './MegaMenuHeader';
 import { MegaMenuItem } from './MegaMenuItem';
 import { usePinnedItems } from './hooks';
@@ -128,11 +125,6 @@ export const MegaMenu = memo(
               ))}
             </ul>
           </ScrollContainer>
-          {shouldRenderInviteUserButton && (
-            <div className={styles.inviteNewUserButton}>
-              <InviteUserButton />
-            </div>
-          )}
         </nav>
       </div>
     );
@@ -169,11 +161,6 @@ const getStyles = (theme: GrafanaTheme2) => {
       [theme.breakpoints.up('md')]: {
         width: MENU_WIDTH,
       },
-    }),
-    inviteNewUserButton: css({
-      display: 'flex',
-      padding: theme.spacing(1.5, 1, 1.5, 1),
-      borderTop: `1px solid ${theme.colors.border.weak}`,
     }),
     dockMenuButton: css({
       display: 'none',
