+++ public/app/features/commandPalette/actions/staticActions.ts
@@ -3,7 +3,10 @@ import { useMemo } from 'react';
 import { NavModelItem } from '@grafana/data';
 import { t } from '@grafana/i18n';
 import { enrichHelpItem } from 'app/core/components/AppChrome/MegaMenu/utils';
-import { performInviteUserClick, shouldRenderInviteUserButton } from 'app/core/components/InviteUserButton/utils';
+import {
+  shouldRenderInviteUserButton,
+  performInviteUserClick,
+} from 'app/core/components/AppChrome/TopBar/InviteUserButtonUtils';
 import { changeTheme } from 'app/core/services/theme';
 import { currentMockApiState, toggleMockApiAndReload, togglePseudoLocale } from 'app/dev-utils';
 import { useSelector } from 'app/types/store';
